Add support for positional + stdin placeholders
This is a breaking change, changing the behavior of the `${command}`
placeholder syntax which previously expanded into an environment
variable called `command`. This syntax is now rather converted into
positional argument(s). The environment variable syntax has been changed
to `={command}`. A third syntax has been created for `stdin` expansion,
piping the output of the placeholder argument into the `stdin` of the
calling command. It's syntax is `<{command}`.

 - `${command}`: Expand `command` as positional argument.
 - `={command}`: Expand `command` as an environment variable called `command`.
 - `<{command}`: Expand `command` as `stdin` for the called command.